add models.py file
